refactor: move shared infrastructure to lib/core/
Moves app_died_exception.dart, launch_failure_analyzer.dart,
process_utils.dart, vm_lifecycle_events.dart, vm_service.dart from
lib/ root to lib/core/ as the first step of the layered architecture.

Updates all 33 importing files. No behavior change; analyze + tests
still pass.

Refs: fdb-grf
