+++ lib/commands/back.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
-import 'package:fdb/app_died_exception.dart';
-import 'package:fdb/vm_service.dart';
+import 'package:fdb/core/app_died_exception.dart';
+import 'package:fdb/core/vm_service.dart';
 
 /// Triggers Navigator.maybePop() in the running Flutter app.
 ///
